Moved bios, cpu and enclosure under info folder
