+++ plugins/module_utils/idrac_utils/info/controller_enclosure.py
