+++ plugins/module_utils/idrac_utils/info/cpu.py
