+++ plugins/modules/idrac_system_info.py
@@ -91,9 +91,9 @@ error_info:
 import json
 from ansible_collections.dellemc.openmanage.plugins.module_utils.dellemc_idrac import iDRACConnection, idrac_auth_params
 from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.firmware.info import IDRACFirmwareInfo
-from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.bios.info import IDRACBiosInfo
-from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.cpu.info import IDRACCpuInfo
-from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.controller.enclosure_info import IDRACEnclosureInfo
+from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.info.bios import IDRACBiosInfo
+from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.info.cpu import IDRACCpuInfo
+from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.info.controller_enclosure import IDRACEnclosureInfo
 from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.fan.info import IDRACFanInfo
 from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.nic.info import IDRACNICInfo
 from ansible_collections.dellemc.openmanage.plugins.module_utils.idrac_utils.sensors_battery.info import IDRACSensorsBatteryInfo
